Remove global allocation feature
The feature was a failed attempt to reduce the number
of contacts for students during the pandemic.

It was developed in a rush and doesn't even have proper
tests. The results were effectively garbage and therefore
discarded during the actual assignment.

If someone ever wants to build this again, feel free, but
test it throughly. I was unable to translate it into a
fitting graph theory problem back than. Admittedly at
that point I've never had contact to graph theory. Good
luck!